Intégration de la date de modification et de création dans Biblio, strati, récolte, localisation
