+++ src/AppBundle/Controller/BackendController.php
@@ -56,71 +56,6 @@ class BackendController extends Controller
     }
 
     /**
-     * @Route("{collectionCode}/searchDiff/", name="searchDiff", options={"expose"=true})
-     * @param string $collectionCode
-     * @return Response
-     */
-    public function searchDiffAction($collectionCode)
-    {
-        $collection = $this->getDoctrine()->getManager()
-            ->getRepository('AppBundle:Collection')->findOneBy(['collectioncode' => $collectionCode]);
-
-        $institutionCode = $this->getUser()->getInstitutionCode();
-        $diffManager = $this->get('diff.manager');
-        $diffManager->init($collection);
-
-        $diffComputer = $this->get('diff.computer');
-        $diffComputer->setCollection($collection);
-
-        $diffHandler = new DiffHandler($this->getParameter('export_path').'/'.$institutionCode);
-        $diffHandler->setCollectionCode($collectionCode);
-
-        $response = new StreamedResponse();
-
-        $response->headers->set('Content-Type', 'text/event-stream');
-        $response->headers->set('Cache-Control', 'no-cache');
-
-
-        $this->searchDiffSetCallBack($response, $diffManager, $diffComputer, $diffHandler);
-
-        $response->send();
-    }
-
-    /**
-     * @param StreamedResponse $response
-     * @param DiffManager      $diffManager
-     * @param DiffComputer     $diffComputer
-     * @param DiffHandler      $diffHandler
-     */
-    private function searchDiffSetCallBack($response, $diffManager, $diffComputer, $diffHandler)
-    {
-        $response->setCallback(function() use ($diffManager, $diffComputer, $diffHandler) {
-            $server = new RecolnatServer();
-            $catalogNumbers = [];
-
-            // Nb total d'étapes :  Search / Compute pour chaque entité
-            // +1 étape sauvegarde
-            $server->steps->send(count($diffManager::ENTITIES_NAME) * 2 + 1);
-            $countStep = 0;
-            foreach ($diffManager::ENTITIES_NAME as $entityName) {
-
-                $server->step->send(json_encode(['count' => $countStep++, 'step' => $entityName.' : recherche']));
-                $catalogNumbers[$entityName] = $diffManager->getDiff($entityName);
-                $server->step->send(json_encode(['count' => $countStep++, 'step' => $entityName.' : traitement']));
-
-                $diffComputer->setCatalogNumbers($catalogNumbers);
-                $diffComputer->computeClassname($entityName);
-            }
-
-            $datas = $diffComputer->getAllDatas();
-
-            $server->step->send(json_encode(['count' => $countStep++, 'step' => 'save']));
-            $diffHandler->saveDiffs($datas);
-            $server->step->send(json_encode(['count' => $countStep++, 'step' => 'done']));
-        });
-    }
-
-    /**
      * @Route("/download/{path}", name="download", options={"expose"=true})
      * @param string $path
      * @return Response
