+++ src/AppBundle/Entity/Repository/Abstracts/AbstractRecolteRepository.php
@@ -1,0 +1,105 @@
+<?php
+
+namespace AppBundle\Entity\Repository\Abstracts;
+
+use AppBundle\Entity\Collection;
+use Doctrine\ORM\AbstractQuery;
+
+abstract class AbstractRecolteRepository extends AbstractRecolnatRepository
+{
+
+    /**
+     * @param string $id
+     * @param int    $fetchMode
+     * @return mixed
+     * @throws \Doctrine\ORM\NonUniqueResultException
+     */
+    public function findOneById($id, $fetchMode = AbstractQuery::HYDRATE_OBJECT)
+    {
+        return $this->getQueryFindOneById('recolte', $id)->getOneOrNullResult($fetchMode);
+    }
+
+    /**
+     * @param Collection $collection
+     * @param array      $catalogNumbers
+     * @param int        $hydratationMode
+     * @return array
+     */
+    public function findByCatalogNumbers(
+        Collection $collection,
+        $catalogNumbers,
+        $hydratationMode = AbstractQuery::HYDRATE_ARRAY
+    ) {
+        $qb = $this->createQueryBuilder('r');
+
+        $qb
+            ->select('r')
+            ->addSelect($this->getExprCatalogNumber().' as catalognumber')
+            ->join('r.specimen', 's');
+        $this->setSpecimenCodesWhereClause($collection, $qb, $catalogNumbers);
+
+        return $this->orderResultSetByCatalogNumber($qb->getQuery()->getResult($hydratationMode), 'eventid');
+    }
+
+    /**
+     * @param Collection $collection
+     * @return \Doctrine\ORM\QueryBuilder
+     */
+    public function getQueryBuilderFindByCollection(Collection $collection)
+    {
+        return $this->getEntityManager()->createQueryBuilder()
+            ->select('r.eventid as id')
+            ->from('recolte', 'r')
+            ->join('r.specimen', 's')
+            ->andWhere('s.collection = :collection')
+            ->setParameter('collection', $collection);
+    }
+
+    /**
+     *
+     * @param array $ids
+     * @return array
+     */
+    public function findById($ids)
+    {
+        $qb = $this->getEntityManager()->createQueryBuilder();
+        $qb->select('r')
+            ->from('recolte', 'r', 'r.eventid')
+            ->andWhere($qb->expr()->in('r.eventid', $ids));
+        $qb->setParameter('ids', $ids, 'rawid');
+
+        return $qb->getQuery()->getResult();
+    }
+
+
+    /**
+     * @param Collection $collection
+     * @param array      $catalogNumbers
+     * @return array
+     */
+    public function findByCatalogNumbersUnordered(Collection $collection, $catalogNumbers)
+    {
+        $qb = $this->getEntityManager()->createQueryBuilder()
+            ->select('r')
+            ->from('recoltee', 'r')
+            ->join('specimen', 's')
+            ->andWhere('s.recolte = r.eventid');
+        $this->setSpecimenCodesWhereClause($collection, $qb, $catalogNumbers);
+
+        return $qb->getQuery()->getResult();
+    }
+    /**
+     * @param array  $datas
+     * @param string $id
+     * @return mixed
+     */
+    public function update(array $datas, $id)
+    {
+        $qb = $this->createUpdateQuery($datas);
+
+        $qb->where('a.eventid = HEXTORAW(:id)')
+            ->setParameter('id', $id);
+
+        return $qb->getQuery()->execute();
+    }
+}
