+++ src/AppBundle/Entity/Repository/LocalisationBufferRepository.php
@@ -4,8 +4,7 @@ namespace AppBundle\Entity\Repository;
 
 use AppBundle\Entity\Repository\Abstracts\AbstractLocalisationRepository;
 
-
-class LocalisationRepository extends AbstractLocalisationRepository
+class LocalisationBufferRepository extends AbstractLocalisationRepository
 {
 
 }
