+++ src/AppBundle/Entity/Repository/RecolteBufferRepository.php
@@ -4,7 +4,7 @@ namespace AppBundle\Entity\Repository;
 
 use AppBundle\Entity\Repository\Abstracts\AbstractRecolteRepository;
 
-class RecolteRepository extends AbstractRecolteRepository
+class RecolteBufferRepository extends AbstractRecolteRepository
 {
 
 }
