+++ src/AppBundle/Entity/Repository/SpecimenBufferRepository.php
@@ -5,7 +5,7 @@ namespace AppBundle\Entity\Repository;
 use AppBundle\Entity\Repository\Abstracts\AbstractSpecimenRepository;
 
 
-class SpecimenRepository extends AbstractSpecimenRepository
+class SpecimenBufferRepository extends AbstractSpecimenRepository
 {
 
 }
