+++ src/AppBundle/Entity/Repository/StratigraphyBufferRepository.php
@@ -4,9 +4,6 @@ namespace AppBundle\Entity\Repository;
 
 use AppBundle\Entity\Repository\Abstracts\AbstractStratigraphyRepository;
 
-/**
- * StratigraphyRepository
- */
-class StratigraphyRepository extends AbstractStratigraphyRepository
+class StratigraphyBufferRepository extends AbstractStratigraphyRepository
 {
 }
